+++ lib/core/constants/images.dart
@@ -1,3 +1,5 @@
 class AppImages{
   static const logoGIF = 'assets/niveus.gif';
+  static const re = 'assets/re.png';
+  static const dio = 'assets/dio.png';
 }